api/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
bcrypt==3.2.0
bleach==4.1.0
Brotli==1.0.9
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.3
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.0
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-api-logger==1.0.8
drf-extra-fields==3.1.1
drf-yasg==1.21.3
fonttools==4.29.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.0.3
packaging==21.3
pdfkit==1.0.0
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
pydyf==0.1.2
PyJWT==2.3.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyphen==0.12.0
python-bidi==0.4.2
python-decouple==3.4
python-dotenv==0.20.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.2
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
zopfli==0.1.9
